gui/requirements.txt
pyserial==3.5
ttkbootstrap==1.10.1
Pillow==9.5.0